requirements.txt
gradio>=5.0.0
requests
python-dotenv












